requirements.txt
streamlit
opencv-python
numpy
mysql-connector-python
pandas
gTTS
pygame
deepface
plotly
